Changed layer_id: usize to z_value: f32 for utils functions. If these do end up being publically accessable functions, it probably makes more sense to name them after how the function uses them. Furthermore, supporting auto-tile layers with multiple tiles in a single cell will eventually make the layer_id and the z_value diverge

+++ src/systems.rs
@@ -236,7 +236,7 @@ fn spawn_level(
                             entity_instance,
                             &entity_definition_map,
                             level.px_hei,
-                            layer_id,
+                            layer_id as f32,
                         );
 
                         let mut entity_commands =
@@ -368,7 +368,7 @@ fn spawn_level(
                             let transform = calculate_transform_from_tile_pos(
                                 tile_pos,
                                 layer_instance.grid_size,
-                                layer_id,
+                                layer_id as f32,
                             );
                             commands
                                 .entity(tile_entity)
